refactor: user role타입 디렉토리 변경
