+++ src/admins/admins.service.ts
@@ -11,7 +11,7 @@ import { Repository } from 'typeorm';
 import bcrypt from 'bcrypt';
 import { CreateAdminDto } from './dto/create-admin.dto';
 import { LoginAdminDto } from './dto/login-admin.dto';
-import { ROLE } from 'src/constant/role';
+import { ROLE } from 'src/common/constant/role';
 
 @Injectable()
 export class AdminsService {
